+++ cli/src/commands/mod.rs
@@ -1,5 +1,4 @@
 mod app;
-mod check;
 mod env;
 mod preset;
 mod self_install;
@@ -7,7 +6,6 @@ mod shell;
 mod sys;
 
 pub(crate) use app::*;
-pub(crate) use check::*;
 pub(crate) use env::*;
 pub(crate) use preset::*;
 pub(crate) use self_install::*;